+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,7 @@
 ## New features: 
   1. add methods `size_vars` and `array_vars` for `ROOT.TTree`
+  1. small adjustment for `tree_reduce`
+  1. remove warnings for `splot` 
 
 ## Backward incompatible: